pyproject.toml
[project]
name = "jurisflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno>=2.4.8",
    "openai>=2.17.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
]
